requirements.txt
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.3
biopython==1.85
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.1
click==8.2.1
filelock==3.18.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lxml==6.0.0
MarkupSafe==3.0.2
narwhals==1.47.0
nltk==3.9.1
numpy==2.2.3
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.31.1
pyarrow==20.0.0
pydeck==0.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.26.0
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
smmap==5.0.2
snscrape==0.7.0.20230622
soupsieve==2.6
streamlit==1.47.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
